requirements.txt
# requirements.txt
requests==2.32.3
pycoingecko==3.1.0
PyYAML==6.0.2
psycopg2-binary==2.9.9
tenacity==9.0.0
python-dateutil==2.9.0.post0
